Create packaging script
