+++ internal/eslint-config/index.js
@@ -228,19 +228,5 @@ module.exports = {
 				],
 			},
 		},
-
-		{
-			files: ['**/scripts/**'],
-			rules: {
-				// allow console in scripts
-				'no-console': 'off',
-
-				// allow dev dependencies in scripts
-				'import/no-extraneous-dependencies': [
-					'error',
-					{ devDependencies: true, optionalDependencies: false, peerDependencies: false },
-				],
-			},
-		},
 	],
 };
