--- internal/scripts/.prettierignore
@@ -1,2 +1,0 @@
-# Build output
-dist
