+++ internal/scripts/tsconfig.eslint.json
@@ -3,11 +3,9 @@
     "allowJs": true,
     "noEmit": true
   },
-  "extends": "./tsconfig.json",
+  "extends": "@moleculer-graphql/tsconfig/tsconfig.json",
+  "files": ["create-package-json.mjs"],
   "include": [
-    "src/**/*", // source folder
-    "examples/**/*", // examples folder
-    "scripts/**/*", // scripts folder
     "./*.js", // root javascript files
     "./.*.js", // root javascript config files
     "./.*.cjs", // root commonjs config files
