--- internal/scripts/tsconfig.json
@@ -1,3 +1,0 @@
-{
-  "extends": "@moleculer-graphql/tsconfig/tsconfig.json"
-}
